+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Added "Vanished" prefix to the tab list player name when the player is vanished
 - Added "Vanished" prefix to the message that the joining player receives while being vanished
+- Fixed that vanished players need to sleep in order for the night to be skipped
 
 -----------Changelog for v1.1.3.1-----------